Making new version of lib parser.
