+++ src/bin/main.rs
@@ -12,7 +12,7 @@ use std::path::Path;
 use std::fs::OpenOptions;
 use std::io::Write;
 use std::env;
-use renju::file_reader::open_file;
+use renju::file_reader::{open_file, open_file_legacy};
 use renju::board_logic;
 use renju::move_node::{MoveIndex, MoveGraph};
 quick_main!(run);
@@ -27,12 +27,21 @@ fn run() -> Result<i32> {
         .arg(Arg::with_name("output")
                  .short("o")
                  .help("File to output to"))
+        .arg(Arg::with_name("legacy")
+                 .short("l")
+                 .takes_value(false)
+                 .help("Parse lib with legacy code"))
         .get_matches();
 
     let path = Path::new(matches.value_of("file").unwrap());
     println!("File: {:?}\n", path);
-    let graph = open_file(&path)
-        .chain_err(|| format!("while parsing file {:?}", path))?;
+    let graph = if !matches.is_present("legacy") {
+        open_file(&path)
+            .chain_err(|| format!("while parsing file {:?}", path))?
+    } else {
+        open_file_legacy(&path)
+            .chain_err(|| format!("while parsing file {:?} in lagacy mode", path))?
+    };
     println!("{:?}", graph);
     //let mut file = OpenOptions::new().write(true).create(true).open(format!("{}.dot",path.file_stem().unwrap().to_str().unwrap())).expect("Couldn't create .dot file");
     //write!(file, "{:?}", graph).chain_err(|| "while writing to file");
@@ -43,10 +52,14 @@ fn run() -> Result<i32> {
         //println!("{:?}", read);
         match read {
             Ok(ref empty) if &empty == &"" => {
-                println!("Exit with ctrl+d");
+                println!("Exit with quit/q or ctrl+d");
             }
             Ok(ref g) if &g == &"graph" || &g == &"g" => {
                 println!("{:?}", graph);
+            }
+            // Should be regex or match, quiz should not match
+            Ok(ref quit) if quit.to_lowercase().starts_with("q") => {
+                return Ok(0);
             }
             Ok(line) => {
                 match line.parse() {
